Make expectations in stepping tests more flexible
We can now write expectations in the following form:
// EXPECTATIONS [backends...] [frontends...]

Backends are specified as in the TargetBackend enum.
Frontends are specified as in the FrontendKinds object.

This will make it more suitable for supporting Kotlin/JS in stepping
tests.

For example, for expectations that match only JVM backends with the
classic frontend, we use:
// EXPECTATIONS JVM JVM_IR ClassicFrontend